+++ back/src/routers/commentRouter.js
@@ -26,6 +26,7 @@ commentRouter.post('/:wantedId', async (req, res, next) => {
         const newComment = { commentContent };
 
         if (!commentContent) {
+            res.status(400).send({error: '모든 값을 입력했는지 확인해주세요.'})
             throw new Error('모든 값을 입력했는지 확인해주세요.');
         }
 
@@ -48,8 +49,15 @@ commentRouter.put('/:commentId', async (req, res, next) => {
         const { commentContent } = req.body;
         const newComment = { commentContent };
 
+        const comment = await commentService.findComment({ commentId })
+        if(!comment){
+            res.status(400).send({error: '이 댓글은 존재하지 않습니다.'})
+            throw new Error(`이 댓글은 존재하지 않습니다.`);
+        }
+
         if (!commentContent) {
-            throw new Error('모든 값을 입력했는지 확인해주세요.');
+            res.status(400).send({error: '모든 값을 입력했는지 확인해주세요.'})
+            throw new Error('모든 값을 입력했는지 확인해주세요.');    
         }
         const updatedComment = await commentService.updateComment({ userId, commentId, newComment });
         console.log(updatedComment)
@@ -64,6 +72,12 @@ commentRouter.delete('/:commentId', async (req, res, next) => {
     try {
         const userId = req.currentUserId;
         const { commentId } = req.params;
+
+        const comment = await commentService.findComment({ commentId })
+        if(!comment){
+            res.status(400).send({error: '이 댓글은 존재하지 않습니다.'})
+            throw new Error(`이 댓글은 존재하지 않습니다.`);
+        }
 
         const deletedComment = await commentService.deleteComment({ userId, commentId });
